+++ src/cli/flag/min_length.rs
@@ -36,7 +36,7 @@ impl FlagHelp for MinLength {
 
 impl FlagInput for MinLength {
     fn produce_input_setting(&self, value: &str) -> Result<Setting, ArgumentError> {
-        let length = value.parse::<u32>().ok().ok_or(INVALID_INPUT_ERROR)?;
+        let length = value.parse::<usize>().ok().ok_or(INVALID_INPUT_ERROR)?;
         Ok(Setting::MinLength(length))
     }
 }
